refactor(parse-tmx, xml2js): file-organization sweep over both packages
parse-tmx's parsers read the models and reach for xml2js and data-urls,
which makes them services rather than utilities, so `src/util` is
`src/services`. xml2js's four private helpers (the empty check, the
processor chain, the pollution-safe property definer and the nullish
entries read) are files of their own; the barrel now carries them, so
both bundle-size snapshots move.

The ledger row is split: vue-phaserjs is three units of its own.
