+++ packages/parse-tmx/src/services/checkIsExternalTileset.ts
