+++ packages/parse-tmx/src/services/cloneNodeWithType.ts
