+++ packages/parse-tmx/src/services/parseFlips.ts
