+++ packages/parse-tmx/src/services/parseObjectShape.ts
