+++ packages/parse-tmx/src/services/parseProperties.ts
