+++ packages/parse-tmx/src/services/parseXmlString.ts
@@ -1,4 +1,4 @@
-import { parseXmlValue } from "#src/util/parseXmlValue";
+import { parseXmlValue } from "#src/services/parseXmlValue";
 import { parseStringPromise } from "@esposter/xml2js";
 import parseDataUrl from "data-urls";
