+++ packages/parse-tmx/src/services/parseXmlValue.test.ts
@@ -1,4 +1,4 @@
-import { parseXmlValue } from "#src/util/parseXmlValue";
+import { parseXmlValue } from "#src/services/parseXmlValue";
 import { describe, expect, test } from "vitest";
 
 describe(parseXmlValue, () => {
