+++ packages/parse-tmx/src/services/parseXmlValue.ts
