+++ packages/xml2js/src/checkIsEmpty.ts
@@ -1,0 +1,2 @@
+export const checkIsEmpty = (value: unknown): boolean =>
+  typeof value === "object" && value !== null && Object.keys(value).length === 0;
