+++ packages/xml2js/src/defineProperty.ts
@@ -1,0 +1,9 @@
+export const defineProperty = (object: Record<string, unknown>, key: string, value: unknown): void => {
+  // Make sure the descriptor hasn't been prototype polluted
+  const descriptor = Object.create(null);
+  descriptor.value = value;
+  descriptor.writable = true;
+  descriptor.enumerable = true;
+  descriptor.configurable = true;
+  Object.defineProperty(object, key, descriptor);
+};
